+++ src/app/anders/page.tsx
@@ -4,6 +4,6 @@ import Link from "next/link";
 export default () => (
     <div className="pt-5 pl-5">
         <Link className="border-gray-500 border p-2" href='/'>Tilbage</Link>
-        <div className="flex min-h-screen flex-col items-center p-24">Lars playground</div>
+        <div className="flex min-h-screen flex-col items-center p-24">Anders playground</div>
     </div>
 )
